pyproject.toml
[project]
name = "dictate"
version = "0.1.0"
description = "Local real-time speech dictation tool for macOS using MLX"
requires-python = ">=3.11"
dependencies = [
    "parakeet-mlx>=0.1.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12",
    "numpy>=1.26",
    "pyperclip>=1.8",
    "pynput>=1.7",
    "tomli>=2.0",
    "pyobjc-framework-Quartz>=10.0",
]

[project.optional-dependencies]
whisper = ["mlx-audio>=0.1.0"]

[project.scripts]
dictate = "dictate.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dictate"]
